fixed id types in repositories

+++ src/main/java/com/uniadmission/universityadmissions/repositories/AdmissionRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdmissionRepository extends JpaRepository<AdmissionEntity, Integer> {
+public interface AdmissionRepository extends JpaRepository<AdmissionEntity, Long> {
 
 	AdmissionEntity getByAdmissionID(Long Id);
